README, more code reorganizing, and no more ref usage
Added a readme to describe installation and play of the application.
Moved the header to `index.html` because it is an always present constant, so no need to dynamically render it.
Formatted more code, rewrote unnecessary code.
Inputs no longer rely on refs to get the input values, but use the returned event object instead.