Pause functionality
Added a button to go to a pause screen and another to resume the game. Had to tweak the timer logic a bit so that it resumes when the game resumes and doesn't count time while the game is paused. Also added a callback so that the game autopauses when the screen loses focus.